Should really use `csize_t` not `uint` since `posix.IOVec` uses that.

+++ cligen/osUt.nim
@@ -404,15 +404,15 @@ proc wr0term*(fd: cint, buf: string): int =
 proc wrLine*(fd: cint, buf: string): int =
   ## Write `buf` & then a single newline atomically (`writev` on Linux).
   let nl = '\n'
-  let iov = [ IOVec(iov_base: buf[0].unsafeAddr, iov_len: buf.len.uint),
+  let iov = [ IOVec(iov_base: buf[0].unsafeAddr, iov_len: buf.len.csize_t),
               IOVec(iov_base: nl.unsafeAddr    , iov_len: 1) ]
   writev(fd, iov[0].unsafeAddr, 2)
 
 proc wrLenBuf*(fd: cint, buf: string): int =
   ## Write `int` length prefix & `buf` data atomically (`writev` on Linux).
   let n = buf.len
-  let iov = [ IOVec(iov_base: n.unsafeAddr     , iov_len: n.sizeof.uint),
-              IOVec(iov_base: buf[0].unsafeAddr, iov_len: buf.len.uint) ]
+  let iov = [ IOVec(iov_base: n.unsafeAddr     , iov_len: n.sizeof.csize_t),
+              IOVec(iov_base: buf[0].unsafeAddr, iov_len: buf.len.csize_t) ]
   writev(fd, iov[0].unsafeAddr, 2)
 
 proc lgBold*(f: File, s: string) = f.write "\e[1m", s, "\e[22m"
